android: frontend: Start a service to keep the persistent notification
